refactor core
