+++ bgmi/downloader/transmission.py
@@ -6,10 +6,8 @@ from bgmi.config import (
     TRANSMISSION_RPC_URL,
     TRANSMISSION_RPC_USERNAME,
 )
-from bgmi.downloader.base import BaseDownloadService
 from bgmi.plugin.base import BaseDownloadService
 from bgmi.plugin.status import DownloadStatus
-from bgmi.website.model import Episode
 
 
 class TransmissionRPC(BaseDownloadService):
@@ -21,10 +19,8 @@ class TransmissionRPC(BaseDownloadService):
             password=TRANSMISSION_RPC_PASSWORD,
         )
 
-    def add_download(self, episode: Episode, save_path: str, overwrite: bool = False):
-        torrent = self.client.add_torrent(
-            episode.download, download_dir=save_path, paused=False
-        )
+    def add_download(self, url: str, save_path: str, overwrite: bool = False):
+        torrent = self.client.add_torrent(url, download_dir=save_path, paused=False)
         return torrent.hashString
 
     @staticmethod
@@ -34,11 +30,11 @@ class TransmissionRPC(BaseDownloadService):
     def get_status(self, id: str) -> DownloadStatus:
         torrent = self.client.get_torrent(id)
         if torrent.error:
-            return DownloadStatus.error
+            return DownloadStatus.not_found
         return {
             "check pending": DownloadStatus.downloading,
             "checking": DownloadStatus.downloading,
             "downloading": DownloadStatus.downloading,
             "seeding": DownloadStatus.done,
             "stopped": DownloadStatus.not_downloading,
-        }.get(torrent.status)
+        }.get(torrent.status, DownloadStatus.error)
