+++ pyproject.toml
@@ -45,6 +45,7 @@ pydantic = "==1.8.2"
 # RPC library
 transmission-rpc = "==3.2.6"
 qbittorrent-api = "==2021.5.22"
+bencoding = "^0.2.6"
 
 [tool.poetry.dev-dependencies]
 black = { version = "21.6b0", python = '^3.6.2' }
@@ -136,5 +137,7 @@ disable = [
   'too-many-statements',
   'too-many-arguments',
   'no-self-use',
-  'not-an-iterable'
+  'not-an-iterable',
+  'redefined-builtin',
+  'super-init-not-called'
 ]
